+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0](https://https://github.com/chrispahm/geos-wasm/compare/v3.0.0...v3.0.1) - 2024-12-05
+
+### Added
+
+- Support for z-values in GeoJSON helper functions
+
+## [3.0.2](https://https://github.com/chrispahm/geos-wasm/compare/v3.0.0...v3.0.1) - 2024-11-26
+
+### Fixed
+
+- fixed Emscripten Module types
+
+## [3.0.1](https://https://github.com/chrispahm/geos-wasm/compare/v3.0.0...v3.0.1) - 2024-11-25
+
+### Fixed
+
+- fixed GEOSversion() function signature
+
 ## [3.0.0](https://https://github.com/chrispahm/geos-wasm/compare/v2.0.0...v3.0.0) - 2024-11-25
 
 ### Breaking Changes